requirements.txt
pvporcupine
sounddevice
webrtcvad
faster-whisper
numpy
python-dotenv
dateparser
feedparser
soundfile